+++ cg/cli/base.py
@@ -23,8 +23,8 @@ from .set import set_cmd
 from .status import status
 from .store import store
 from .transfer import transfer
-from .upload import upload
-from cg.cli.trending import trending as trending_command
+from .upload.upload import upload
+from .upload import vogue as vogue_command
 
 
 LOG = logging.getLogger(__name__)
@@ -83,5 +83,5 @@ base.add_command(transfer)
 base.add_command(upload)
 analysis.add_command(balsamic)
 analysis.add_command(rna)
-upload.add_command(trending_command)
+upload.add_command(vogue_command)
