+++ cg/cli/upload/__init__.py
@@ -1,2 +1,2 @@
 """Init for the trending CLI"""
-from .base import trending
+from .vogue import vogue
