+++ cg/meta/upload/vogue.py
@@ -15,12 +15,14 @@ class UploadVogueAPI():
 
     def load_genotype(self, days):
         """Loading genotype data from the genotype database into the trending database"""
-        samples = self.genotype_api.prepare_sample(days=days)
+        samples = self.genotype_api.export_sample(days=days)
         samples = json.loads(samples)
-        for sample_id in samples:
-            self.vogue_api.load_genotype(samples[sample_id])
+        for sample_id, sample_dict in samples.items():
+            sample_dict['_id'] = sample_id
+            self.vogue_api.load_genotype_data(sample_dict)
         
-        samples_analysis = self.genotype_api.prepare_analysis(days=days)
+        samples_analysis = self.genotype_api.export_sample_analysis(days=days)
         samples_analysis = json.loads(samples_analysis)
-        for sample_id in samples_analysis:
-            self.vogue_api.load_genotype(samples_analysis[sample_id])
+        for sample_id, sample_dict in samples_analysis.items():
+            sample_dict['_id'] = sample_id
+            self.vogue_api.load_genotype_data(sample_dict)
